Add support for the / prefix.
On Windows command line arguments could be prefixed with a / character.
On Linux this would be a problem since that is the directory path seperator (
on Windows its \). This change set updates the code to only add the / character
as a valid argument prefix _if_ that is not also the systems directory path seperator.
Basically means this prefix will work on Windows but not on Linux.