Add copy and move constructors/assignment operators. To make the returning of a copy of the builder more effecient I've added a move constructor and a move assignment operator.
Noticed a copy constructor and assignment operator was in order so add them too.

Remove the function call operator which returned a pointer to a CommandLineInferface as that would just risk memory leak or
sigfault issues. Now you can only get a copy of the CLI from the builder and the builder will delete its copy in
its destructor.