Add support for multipart arguments. That is arguments which are of a key-value pair and use a space to seperate the two. In this case the key and value come in as two seperate arguments and need to be viewed together as a multi part argument.
I followed TDD's Red-Green-Refactor process and this commit is the Red-Green part. I still
need to refactor as its working but its rough, just kind of slapped together and needs to
be cleaned up. I'm commit it now because at least its Green and I don't want to fix my self
to death.