Add support for default values. Now users can define a value argument and provide a default value to be used if the argument is not provided on the CLI.
So you can define an argument called 'address' and set a default vlaue to
'127.0.0.1'. In your code you can just read 'commandLinearg[address]' and
you will get either 127.0.0.1 if the user does not use the address argument
or what ever value the user inputted via the address argument. With this additoin
you no longer need to get to see if the address argument was used before defaulting
the value used, you can just use the value stroed in the argument and trust that its
either user defined or your default.