Fix Doxygen renderings of bold and italic text in mark down. Doxygen's handling of text emphasis is limited compared to an actual Markdown engine. Where as GitHub was fine with *--help* Doxygen didn't like it because the text did not start with an alphanumeric value (i.e. it started with a dash dash). To fix this limitation in Doxygen I'm inserting mark up to emphasis all the text which start with a non-alphanumeric character.
Also for the same reason Doxygen does not seem to support the combination of bold and italic on
the same text in the markdown. This would be the tag line 'Why Reinvent the Wheel? Just use Taranis!'
To make this text properly stand out I'm turning it into a h2 header. That way we can keep it as markdown
v.s. markup. Also by making it an h2 it stands out more then if it was just bold & italic and the following
paragraph does fit with the tag line as its header.