Fix ambiguoity issue in Visual Studio 2013. In Visual Studio 2013 (which is C++0x) the std::function feature seems to be only partially supported where in that if you pass it a char* it will implicity bind to it.
The result is that the WithValue() method has two overloads which takes three arguments where the
last argument is either a QString or a std::function. When using said method when you put
in three string literals what you are doing is putting in three char* which get implicitly
converted into QString's. However a char* can also be implicitly converted into a std::function
thus the ambiguouity.

In Visual Studio 2015 (C++11) this is not an issue.

To deail with this I'm explicilty wrapping the string liternals in the QStringLieral macro
which will create a QString object instead of a char* thus resolving the ambiguouity and choosing
the right overload.